pyproject.toml
[project]
name = "dbtt"
version = "0.1.2"
description = "A toolkit designed to enhance and streamline your dbt workflows."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Cristian Webber" }

]
dependencies = [
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "typer>=0.15.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.8.6",
]

[project.scripts]
dbtt = "cli:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
